factcheckmodal UI update
